Remove OnTimerInvokers.INSTANCE; deprecate DoFnInvokers.INSTANCE
